Update shell_script.sh
New Features Added:
a.It is not necessary to have only Ashburn and Phoenix Region now for tenancies. New property has been added in ocswork.properties and VCN-Info tab in CD3 for mentioning the Regions your tenancy is subscribed to and based on that directories and other values in variables file will be populated.
b.Add new Subnet/DHCP. When Networking objects have been created and modified as per requirements(may be few instances also created) and then after that there is requirement to add a new subnet or DHCP then instead of re-creating the existing network; you can chose to just add new subnet and it will keep other networking things as is.